refactor: split test_sync_executor_service.py by concern
Splits the 1057-line test_sync_executor_service.py into two files along
the boundary identified in its own module docstring:

- test_sync_executor_service_wiring.py: executor construction,
  wire_services() registration, dependency-graph validation, and the
  AppSync shutdown-ordering regression.
- test_sync_executor_service_saturation.py: saturation-warning behavior,
  the periodic probe, shutdown interruption of busy-loop/C-blocked
  workers, and config-driven behavior (max_workers/shutdown_timeout).

Shared test factory helpers (make_test_config, make_sync_executor_hassette,
make_service) are duplicated in both files, matching the existing pattern
in tests/unit/core/test_command_executor*.py.

No test was dropped — both files together contain the same 51 test
functions as the original, verified by diffing the sorted test-name sets.

Also updates a stale cross-reference in
tests/unit/task_bucket/test_interruptible_executor.py that named the old
file directly.

Closes #1262
